Changed ptrdiff_t to int in property accessors.  This is for compatibility with some changes in clang that are required to prevent things from breaking on LP64 platforms.
It would be cleaner to use a ptrdiff_t here, but unfortunately we can't because the ivar_offset field in the runtime metadata is an int so ivar offsets beyond 4GB will not work on LP64 platforms.

This probably isn't a limitation.  If you have more than 4GB of ivars in one object, you've done something badly wrong and probably shouldn't be allowed to write code anymore.

+++ properties.m
@@ -58,7 +58,7 @@ inline static void lock_spinlock(int *spinlock)
 	}
 }
 
-id objc_getProperty(id obj, SEL _cmd, ptrdiff_t offset, BOOL isAtomic)
+id objc_getProperty(id obj, SEL _cmd, int offset, BOOL isAtomic)
 {
 	char *addr = (char*)obj;
 	addr += offset;
@@ -79,7 +79,7 @@ id objc_getProperty(id obj, SEL _cmd, ptrdiff_t offset, BOOL isAtomic)
 	return [ret autorelease];
 }
 
-void objc_setProperty(id obj, SEL _cmd, ptrdiff_t offset, id arg, BOOL isAtomic, BOOL isCopy)
+void objc_setProperty(id obj, SEL _cmd, int offset, id arg, BOOL isAtomic, BOOL isCopy)
 {
 	if (isCopy)
 	{
